% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster.r
\name{add_cluster}
\alias{add_cluster}
\title{Função auxiliar para adicionar a coluna de cluster na base original.}
\usage{
add_cluster(base, var, n_clusters)
}
\arguments{
\item{base}{Base de dados contendo as colunas TempoCasa, event e var.}

\item{var}{nome da coluna a ser clusterizada.}

\item{n_clusters}{número de clusters.}
}
\value{
base com uma coluna extra chamada de varClusters
}
\description{
Função auxiliar para adicionar a coluna de cluster na base original.
}
