% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster.r
\name{cluster_km}
\alias{cluster_km}
\title{Clusterização via K-Means Longitudinal usando o pacote kml.}
\usage{
cluster_km(base, var, n_clusters)
}
\arguments{
\item{base}{Base de dados contendo as colunas TempoCasa, event e var.}

\item{var}{nome da coluna a ser clusterizada.}

\item{n_clusters}{número de clusters.}
}
\value{
data.frame contendo os valores distintos da coluna original
e o cluster a qual cada um deles pertence.
}
\description{
Clusterização via K-Means Longitudinal usando o pacote kml.
}
