% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_extract.R
\name{completa_surv}
\alias{completa_surv}
\title{Função para preencher os buracos na função de sobrevivência.}
\usage{
completa_surv(sobrev)
}
\arguments{
\item{sobrev}{data frame com as colunas 'time', 'id' e 'wx'}
}
\value{
data.frame com as mesmas colunas que sobrev, mas
sem buracos na coluna time. Os wx's faltantes são preenchidos
por zero.
}
\description{
Função para preencher os buracos na função de sobrevivência.
}
