% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_extract.R
\name{fix_km_names}
\alias{fix_km_names}
\title{Função para consertar a forma como o km faz o nome de cada grupo.}
\usage{
fix_km_names(km_names)
}
\arguments{
\item{km_names}{names do componente "strata" do resultado
de um ajuste km usando survfit.}
}
\value{
vetor com os nomes corrigidos.
}
\description{
Aplica a função para sanitizar os nomes do KM para o caso onde
existem varias variaveis explicativas.
}
