% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_extract.R
\name{fix_km_names_aux}
\alias{fix_km_names_aux}
\title{Função para consertar a forma como o km faz o nome de cada grupo.}
\usage{
fix_km_names_aux(km_names)
}
\arguments{
\item{km_names}{names do componente "strata" do resultado de um
ajuste km usando survfit.}
}
\value{
um vetor com os nomes formatados.
}
\description{
Função para consertar a forma como o km faz o nome de cada grupo.
}
