% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getKMcurve}
\alias{getKMcurve}
\title{Extract data describing the Kaplan-Meier curve from a \code{\link[survival]{survfit}} object}
\usage{
getKMcurve(km)
}
\arguments{
\item{km}{A survfit object containing the Kaplan-Meier estimate of the
survival function as returned by \code{\link[survival]{survfit}}}
}
\value{
A data frame with columns:
  \describe{
    \item{Time}{The time of the events}
    \item{Survival}{The probability of survival at each event time}
    \item{Nrisk}{The number of patients at risk at each event time}
    }
}
\description{
Extract data describing the Kaplan-Meier curve from a \code{\link[survival]{survfit}} object
}

