% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.r
\name{km_aj}
\alias{km_aj}
\title{Ajuste e plot dos resultados do KM}
\usage{
km_aj(base, abertura, time = "TempoCasa", event = "event")
}
\arguments{
\item{base}{base com as colunas abertura, TempoCasa e event.}

\item{abertura}{nome da coluna com a variável explicatória.}

\item{time}{nome da coluna com os tempos até falha/censura.}

\item{event}{nome da coluna com o status de falha/censura.}
}
\value{
Nada
}
\description{
Ajusta o KM usando o pacote survival e plota os resultados.
Plota tanto a curva de sobrevivência, quanto o percentual
de cancelamentos em cada tempo.
}
\examples{
require(survival)
km_aj(lung, 'sex', 'time', 'status')

}
