% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.r
\name{km_aj_cluster}
\alias{km_aj_cluster}
\title{Função para ajustar e plotar o KM com uma variável explicatória,
comparando com a sua clusterização}
\usage{
km_aj_cluster(base, abertura, time = "TempoCasa", event = "event")
}
\arguments{
\item{base}{base com as colunas abertura, TempoCasa e event.}

\item{abertura}{nome da coluna com a variável explicatória}

\item{time}{nome da coluna com os tempos até falha/censura.}

\item{event}{nome da coluna com o status de falha/censura.}
}
\value{
invisible
}
\description{
Função para ajustar e plotar o KM com uma variável explicatória,
comparando com a sua clusterização
}
