% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{surv_aj}
\alias{surv_aj}
\title{Fit a KM estimate with only one covariate}
\usage{
surv_aj(x, data, ...)
}
\arguments{
\item{x}{name of the covariate column}

\item{data}{data.frame with the survival data.}

\item{...}{extra arguments passed to `survival::survfit`}
}
\value{
an object of class "survfit". See survfit.object for details. Methods defined for survfit objects are print, plot, lines, and points.
}
\description{
Wrapper around `survival::survfit.formula` for allowing the use of `purrr::map` and fit
several one covariate KM estimates at a time.
}
\seealso{
\code{\link{survfit.formula}}
}
