% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_cluster.r
\name{surv_cluster}
\alias{surv_cluster}
\title{K-means clusterization for survival data}
\usage{
surv_cluster(
  aj,
  n_clusters,
  nbRedrawing = 1,
  parAlgo = kml::parALGO(saveFreq = Inf, distanceName = "canberra")
)
}
\arguments{
\item{aj}{aj previously fitted model (using `survival::survfit()`).}

\item{n_clusters}{number of clusters, as defined in `kml::kml`.}

\item{parAlgo, nbRedrawing}{argumentos passados para `kml::kml`.}
}
\value{
a data.frame with the resulting groups.
}
\description{
Clusterization of survival data using `kml` and `survival`. This function
identify clusters of data that behaves similarly with regard to the hazard
function.
}
