% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_extract.R
\name{surv_extract_cox}
\alias{surv_extract_cox}
\title{Função para extrair as curvas de sobrevivência a partir de um ajuste cox.}
\usage{
surv_extract_cox(aj, newdata)
}
\arguments{
\item{aj}{objeto com o resutado do ajuste, usualmente o resultado de um coxph.}

\item{newdata}{objeto com os dados onde o ajuste será realizado. para plotar todas as curvas,
basta usar a combinação única de todas as covariáveis categóricas. Para covariáveis contínuas,
é preciso escolher o valor(ou valores) para os quais a curva deve ser plotada (nao testado).}
}
\value{
data.frame com as colunas time, id, surv e wx.
id apresenta as combinações unicas das covariáveis usadas no ajuste.
surv apresenta as probabilidades de sobrevivencia e wx apresenta as razoes
numero de eventos/numeros de expostos.
}
\description{
Função interna que recebe um modelo de cox previamente ajustado e retorna os
resultados em um formato conveniente.
}
