% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_extract.R
\name{surv_extract_km}
\alias{surv_extract_km}
\title{Função para extrair as curvas de sobrevivência a partir de um ajuste km.}
\usage{
surv_extract_km(aj)
}
\arguments{
\item{aj}{objeto com o resutado do ajuste, usualmente o resultado de um
survfit (km)}
}
\value{
data.frame com as colunas time, id, surv e wx.
id apresenta as combinações unicas das covariáveis usadas no ajuste.
surv apresenta as probabilidades de sobrevivencia e wx apresenta as razoes
numero de eventos/numeros de expostos.
}
\description{
Função para extrair as curvas de sobrevivência a partir de um ajuste km.
}
