% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_plot.r
\name{surv_plot}
\alias{surv_plot}
\title{Plot survival curves using survival and plotly package.}
\usage{
surv_plot(aj, type = c("both", "surv", "wx"))
}
\arguments{
\item{aj}{aj previously fitted model (using `survival::survfit()`).}

\item{type}{should the "wx", "survival" or "both" be plotted?}
}
\value{
a plotly with two plots, one for survival probabilities and
another for wx.
}
\description{
Plot a survival K-M estimate and shows both survival and n.event/n.risk.
}
\examples{

require(survival)
aj <- survfit(Surv(time, status) ~ sex, data = lung)
surv_plot(aj)

}
