% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_plot.r
\name{surv_plot_cluster}
\alias{surv_plot_cluster}
\title{Plot a comparison between the unclusterized and clusterized survival.}
\usage{
surv_plot_cluster(
  x,
  data,
  time = "time",
  event = "event",
  x_cluster = paste0(x, "Cluster"),
  ...
)
}
\arguments{
\item{x}{name of the covariate column.}

\item{data}{data.frame with the survival data.}

\item{time}{name of the time variable column.}

\item{event}{name of the censor/death variable column.}

\item{x_cluster}{name of the clusterized covariate}

\item{...}{extra arguments passed to `survival::survfit`.}
}
\value{
a plotly graph comparing the clusterization of `x` with regard to
wx.
}
\description{
Plot a comparison between the unclusterized and clusterized survival.
}
\details{
This function uses the following wrapper to fit a KM estimate
based on its arguments, for the clusterized and unclusterized column:

`survfit(formula(paste0("Surv(time,event)~", x)), data = data, ...)`

`survfit(formula(paste0("Surv(time,event)~", x_cluster)), data = data, ...)`
}
