% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{truncate_interval}
\alias{truncate_interval}
\title{Truncate survival times}
\usage{
truncate_interval(data, time_int, startDate, endDate, censor)
}
\arguments{
\item{data}{data.frame}

\item{time_int}{lubridate::interval output, with the time intervel where the survival times will be
truncated}

\item{startDate}{name of the column in data with the birth date of each subject (birth date can be seen as policy begin)}

\item{endDate}{name of the column in data with the death/censor date of each subject}

\item{censor}{name of the column with the censoring status. 1 if endDate is a death date, or 0 if endDate is a censor date.
if endDate is later than interval end, then the end date will be the interval end and the status will be censor.}
}
\value{
a data.table with the extra columns: event, dataInicio, dataFim, tempoInicio, tempoFim
}
\description{
Function to truncate survival times in a time interval, for investiganting survival probabilities changes in time.
}
