% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_cluster.r
\name{add_cluster}
\alias{add_cluster}
\title{Add the clusterization results to the full data}
\usage{
add_cluster(data, cluster_res, id, name_to = paste0(id, "Cluster"))
}
\arguments{
\item{data}{data.frame with the full data to wich the clusterized column will be added.}

\item{cluster_res}{export of the `surv_cluster` function, usually with only one covariate.}

\item{id}{name of the unclusterized column, used in the join.}

\item{name_to}{name of the clusterized column.}
}
\value{
a data.frame with the same columns as the input, plus the cluster column.
}
\description{
Add the clusterization results to the full data
}
