% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_extract.R
\name{surv_extract}
\alias{surv_extract}
\title{Função para extrair as curvas de sobrevivência de um ajuste.}
\usage{
surv_extract(aj, newdata = NULL)
}
\arguments{
\item{aj}{resultado de um modelo de km (usando suvfit) ou de um modelo de cox
(usando coxph), previamente ajustado.}

\item{newdata}{data.frame com os valores das covariaveis para a previsao
do modelo de cox. Só é usado quando `class(aj) == "coxph"`.}
}
\value{
data.frame com as colunas time, id, surv e wx.
id apresenta as combinações unicas das covariáveis usadas no ajuste.
surv apresenta as probabilidades de sobrevivencia e wx apresenta as razoes
numero de eventos/numeros de expostos.
}
\description{
Função para extrair as curvas de sobrevivência de um ajuste.
}
\examples{

require(survival)
aj <- survfit(Surv(time, status) ~ sex, data = lung)
surv_extract(aj)

aj <- coxph(Surv(time, status) ~ sex, data = lung)
surv_extract(aj, newdata = data.frame(sex = 1))

aj <- coxph(Surv(time, status) ~ sex, data = lung)
surv_extract(aj, newdata = data.frame(sex = c(1,2)))

}
